requirements.txt
asgiref==3.8.1
Django==5.0.6
psycopg2==2.9.9
djangorestframework==3.15.1
iniconfig==2.0.0
packaging==24.0
pip==24.0
pluggy==1.5.0
pytest==8.2.0
setuptools==65.5.0
sqlparse==0.5.0
python-dotenv==1.0.1
drf-yasg==1.21.7
faker==25.2.0
django-factory_boy==1.0.0
pytest-django==4.8.0
drf_spectacular==0.27.2